+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 Historic and pre-release versions aren't necessarily included.
 
 
-## [UNRELEASED] - tbc
+## [0.17.1] - 2025-11-16
 
 ### Fixed
 
@@ -281,6 +281,7 @@ Historic and pre-release versions aren't necessarily included.
 - Update dev/test dependencies: ruff
 
 
+[0.17.0]: https://github.com/elliot-100/Spond-classes/compare/v0.17.0...v0.17.1
 [0.17.0]: https://github.com/elliot-100/Spond-classes/compare/v0.16.1...v0.17.0
 [0.16.1]: https://github.com/elliot-100/Spond-classes/compare/v0.16.0...v0.16.1
 [0.16.0]: https://github.com/elliot-100/Spond-classes/compare/v0.15.0...v0.16.0